+++ indicators.py
@@ -111,7 +111,7 @@ def text_search(search_term, trommel_vfeed_output):
 		cve_hit = '"(CVE-\d+-\d+ : .*\.)"'
 		name_hit = re.findall(cve_hit, search_text)
 		for match_hit in name_hit:
-			trommel_vfeed_output.write("Found %s and it has been associated with %s\n" % (search_term, match_hit))
+			trommel_vfeed_output.write("Found %s and it has been associated with %s\n\n" % (search_term, match_hit))
 	#Searches above CVE in Exploit-DB and Metasploit
 	for cve_hit in cve_field:
 		edb = exploitdb_result(cve_hit)
@@ -123,7 +123,7 @@ def text_search(search_term, trommel_vfeed_output):
 			url_match = "http://www.exploit-db.com/exploits/\d{1,8}"
 			urls = re.findall(url_match, edb, re.S)
 			for url_hit in urls:
-				trommel_vfeed_output.write("%s has a known exploit: %s\n" % (cve_hit, url_hit))
+				trommel_vfeed_output.write("%s has a known exploit: %s\n\n" % (cve_hit, url_hit))
 		#Metasploit results
 		if msf is not "null":
 			msf_fname = "metasploit-framework/modules/.*\.rb"
@@ -132,20 +132,20 @@ def text_search(search_term, trommel_vfeed_output):
 			msf_title_match = re.findall(msf_title, msf)
 			for match in msf_fn_match:
 				for match2 in msf_title_match:
-					trommel_vfeed_output.write("%s is associated with the following Metasploit Module: %s - %s\n" % (cve_hit, match2, match))
+					trommel_vfeed_output.write("%s is associated with the following Metasploit Module: %s - %s\n\n" % (cve_hit, match2, match))
 		#Snort results
 		if snort is not "null":
 			snort_sid = 'id": "sid:(.*)'
 			snort_sid_match = re.findall(snort_sid, snort)
 			for match in snort_sid_match:
-				trommel_vfeed_output.write("%s is associated with the Snort sid:%s" % (cve_hit, match))
+				trommel_vfeed_output.write("%s is associated with the Snort sid:%s\n\n" % (cve_hit, match))
 		
 		#Nmap results
 		if nmap is not "null":
 			nmap_script = '"file": "(.*)",'
 			nmap_script_match = re.findall(nmap_script, nmap)
 			for match in nmap_script_match:
-				trommel_vfeed_output.write("%s is associated with the Nmap script: %s" % (cve_hit, match))		
+				trommel_vfeed_output.write("%s is associated with the Nmap script: %s\n\n" % (cve_hit, match))